+++ packages/chat-sdk/src/services/realtime-api.ts
@@ -456,7 +456,6 @@ export class RealtimeAPI implements BaseRealtime {
         ],
       },
       (response) => {
-        console.log({ response });
         callback(response as ServerReaction[]);
       },
     );