+++ CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.2.1 (2021-07-13)
+
+
+### Bug Fixes
+
+* listen to chat channel ([44f8b86](https://github.com/stationfy/arena-chat-sdk/commit/44f8b86d0d6d9ca85c724e652686092459527e8f))
+
+
+
+
+
+# 2.2.0 (2021-07-07)
+
+
+### Features
+
+* bump package version ([076a2a5](https://github.com/stationfy/arena-chat-sdk/commit/076a2a5be8cb7c94c75305a2955640d07d2c017b))
+
+
+
+
+
+# 2.2.0-beta.6 (2021-07-07)
+
+
+### Features
+
+* bump package version ([48dfb3d](https://github.com/stationfy/arena-chat-sdk/commit/48dfb3d76b2b5b9781b0cff40f93e3951093b066))
+
+
+
+
+
+# 2.2.0-beta.5 (2021-07-07)
+
+
+### Features
+
+* add liveblog package ([007b74a](https://github.com/stationfy/arena-chat-sdk/commit/007b74a04560b336042a84107356a222c5b75f41))
+
+
+
+
+
 ## 2.0.30 (2021-05-14)
 
 **Note:** Version bump only for package arena-chat-sdk